move json protocol stuff to seperate file
